+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 1. Support following users. 支持关注用户。
 2. Show public playlists in user space. 新增在神人空间显示公开歌单。
+3. Support login device management. 支持登录设备管理。
 
 ## Enhancements 改进